First Time + Under development banner + Status page
1) Status page
Create a status page `/status`, This page should not be linked from the front end. You need to know the url to be able to visit it. Move the "System Status" from the help page `/help` to this new status page. This status page should also list other stats about the heath of the system.

2) Under development banner
Add a Alert banner to the top of the page. This should say that this website is still in development, the data is cleared regularly while we are working on it (don't submit new data, it will get erased), and this webpage will be moved to a new domain shortly. If they have questions they can contact us support@art.abluestar.com

3) First Time
When a new user visits this website for the first time. Pop up the "first time" model dialog.
In this pop up it describes what this project is trying to do, and how they can help. Then links to the help page for more information. This pop up is only shown once and only the first time they visit. There should be a Close button in the top right corner, or they can click outside the box to close it.