The Fast Add artwork page
The uploaded artwork thumbnail should be larger.
The location location information should be to the right of the thumbnail. The location information should also include a mini map.

The "Add report" button should be changed to "Add photo to Artwork"

The "What would you like to do?" section is highlighted weirdly. Dark blue background with a white text.

Location should preference the GPS data in the photo EXIF data over the device GPS.

If the photo GPS data can't be found. Highlight this as an error, and hightlight that the device GPS is being used instead. We want people to use the photo GPS instead of the Device GPS.